cryptoguard/requirements.txt
flask==3.0.3
flask-sqlalchemy==3.1.1
flask-bcrypt==1.0.1
flask-jwt-extended==4.6.0
flask-cors==4.0.1
scikit-learn==1.5.1
numpy==1.26.4
pandas==2.2.2